services/api/pyproject.toml
[project]
name = "agrosaas-api"
version = "0.1.0"
description = "AgroSaaS Modular Monolith API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "loguru>=0.7.2",
    "celery>=5.4.0",
    "redis>=5.0.3",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "geoalchemy2>=0.14.0",
    "shapely>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.5",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
lint.select = ["E", "W", "F", "I", "B", "C4", "UP", "S", "N"]
lint.ignore = ["S101"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
